[HOTSPOT-181] fix: 사용자 데이터 한도 dataLimit 응답 값은 double, 요청, db 저장은 long으로 변경

+++ src/main/java/hotspot/user/family/controller/response/FamilySubscriptionResponse.java
@@ -12,7 +12,7 @@ public record FamilySubscriptionResponse(
         Long familyId,
         FamilyRole role,
         int priority,
-        int dataLimit
+        double dataLimit
 ) {
     public static FamilySubscriptionResponse from(FamilySubscription familySubscription) {
         return new FamilySubscriptionResponse(
